backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.1.7
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2023.4
waitress==3.0.2
Werkzeug==3.1.3
